hw7

+++ domain/src/main/java/com/example/domain/usecase/DeleteTaskUseCase.kt
@@ -3,7 +3,8 @@ package com.example.domain.usecase
 import com.example.domain.repository.TaskRepository
 
 class DeleteTaskUseCase(private val taskRepository: TaskRepository) {
-    suspend operator fun invoke(taskId: Int) {
+
+    suspend operator fun invoke(taskId: Long) {
         taskRepository.deleteTask(taskId)
     }
 }
